hiva_job/requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
coverage==7.6.4
decouple==0.0.7
Django==5.1.3
django-cors-headers==4.6.0
django-guardian==2.4.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
elastic-transport==8.15.1
guardian==0.2.2
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.4
kombu==5.4.2
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg==3.2.3
psycopg2==2.9.10
PyJWT==2.10.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.0
requests==2.32.3
rest-framework-simplejwt==0.0.2
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
